logging in backend controllers

+++ src/web_app/backend/controllers/ProductController.php
@@ -8,7 +8,6 @@ use common\models\Product;
 use common\models\search\ProductSearch;
 use common\models\Type;
 use Yii;
-use yii\data\ActiveDataProvider;
 use yii\filters\AccessControl;
 use yii\filters\VerbFilter;
 use yii\helpers\ArrayHelper;
@@ -76,15 +75,19 @@ class ProductController extends Controller
                         Yii::$app->session->setFlash('Success', 'Successfully Uploaded Images and Created Product!');
                     }else {
                         Yii::$app->session->setFlash('Error', 'Failed to create files!');
+                        Yii::warning('Failed to create files!', __METHOD__);
                     }
                 } else {
+                    $errors = implode('', $product->getErrors());
                     Yii::$app->session->setFlash('Error', 'Failed to create this product!');
+                    Yii::error('Failed to create this product!' . $errors, __METHOD__);
                 }
                 $transaction->commit();
             }catch (\Exception $e){
                 Yii::$app->session->setFlash('Error', $e->getMessage());
                 $transaction->rollBack();
                 File::rrmdir(Yii::getAlias('@frontend/web/storage/images/'.$product->folder_id));
+                Yii::error($e->getMessage(), __METHOD__);
                 return $this->redirect(['/product/products']);
             }
             return $this->redirect(['/product/products']);
@@ -113,16 +116,21 @@ class ProductController extends Controller
             if ($product->save()) {
                 Yii::$app->session->setFlash('Success', 'Successfully Edited the Product!');
             } else {
-              Yii::$app->session->setFlash('Error', 'Failed to Edit the Product!');
+                $errors = implode('', $product->getErrors());
+                Yii::warning('Failed to Edit the Product! ' . $errors  , __METHOD__);
+                Yii::$app->session->setFlash('Error', 'Failed to Edit the Product!');
             }
-            return $this->redirect('/product/products');
+            return $this->redirect('/products');
+        } else if (Yii::$app->request->isGet) {
+            return $this->renderAjax('edit', [
+                'product' => $product,
+                'types' => $types,
+                'brands' => $brands
+            ]);
         }
 
-        return $this->renderAjax('edit', [
-            'product' => $product,
-            'types' => $types,
-            'brands' => $brands
-        ]);
+        Yii::$app->session->setFlash('Error', 'Something Went Wrong!');
+        return $this->redirect('/products');
     }
 
     public function actionDelete($id): Response
@@ -133,16 +141,24 @@ class ProductController extends Controller
             return $this->redirect('/product/products');
         }
 
-
         try {
             $pathName = Yii::getAlias('@frontend/web/storage/images/'.$product->folder_id);
             if (file_exists($pathName)){
-                File::rrmdir($pathName);
+                if (File::rrmdir($pathName)) {
+                    Yii::$app->session->setFlash('Success', 'Successfully Deleted Product!');
+                    Yii::info('Deleted Product at ' . Yii::$app->formatter->asDate(strtotime('now')) , __METHOD__);
+                    $product->delete();
+                } else {
+                    Yii::$app->session->setFlash('Error', 'Failed to Delete Files! Could not delete ' . $pathName);
+                    Yii::error('Failed to Delete Files! Could not delete ' . $pathName , __METHOD__);
+                }
+            } else {
+                Yii::$app->session->setFlash('Error', 'Failed to Delete Files! Maybe ' . $pathName . ' Does not Exist?');
+                Yii::error('Failed to Delete Files! Maybe ' . $pathName . ' Does not Exist?', __METHOD__);
             }
-            $product->delete();
-            Yii::$app->session->setFlash('Success', 'Successfully Deleted Product!');
         } catch (\Throwable $t) {
             Yii::$app->session->setFlash('Error', 'Failed to Delete Product! ' . $t->getMessage());
+            Yii::error('Failed to Delete Product! ' . $t->getMessage(), __METHOD__);
         }
         return $this->redirect('/product/products');
     }
